Refractor Display function

+++ SmartMenu.Domain/Models/DTO/TemplateCreateDTO.cs
@@ -16,10 +16,10 @@ namespace SmartMenu.Domain.Models.DTO
         public string TemplateDescription { get; set; } = string.Empty;
 
         [Range(40f, 8000f, ErrorMessage = "Template Width must be between 40 and 8000f.")] // Example range constraint
-        public float TemplateWidth { get; set; }
+        public int TemplateWidth { get; set; }
 
         [Range(40f, 3125f, ErrorMessage = "Template Height must be between 40 and 3125f.")] // Example range constraint
-        public float TemplateHeight { get; set; }
+        public int TemplateHeight { get; set; }
 
         public TemplateType TemplateType { get; set; }
